undo the update of the star rating,  if there is an error setting on the server
also console.log the error for debugging

note the spec doesn't specify error behavior.  previously we had been
waiting for the update to complete on the server before updating the
UI, but I was noticing it was a tad slow, and I hate slow.

this could be undone if people don't like it.